# execution, engine: misc fuzzing fixes