--- a/engine/engine.go
+++ b/engine/engine.go
@@ -160,6 +160,13 @@ func New(opts Opts) *compatibilityEngine {
 	}
 }
 
+var (
+	// Duplicate label checking logic uses a bitmap with 64 bits currently.
+	// As long as we use this method we need to have batches that are smaller
+	// then 64 steps.
+	ErrStepsBatchTooLarge = errors.New("'StepsBatch' must be less than 64")
+)
+
 type compatibilityEngine struct {
 	prom      v1.QueryEngine
 	functions map[string]*parser.Function
@@ -208,6 +215,9 @@ func (e *compatibilityEngine) NewInstantQuery(ctx context.Context, q storage.Que
 		ExtLookbackDelta:         e.extLookbackDelta,
 		EnableAnalysis:           e.enableAnalysis,
 		NoStepSubqueryIntervalFn: e.noStepSubqueryIntervalFn,
+	}
+	if qOpts.StepsBatch > 64 {
+		return nil, ErrStepsBatchTooLarge
 	}
 
 	lplan, warns := logicalplan.New(expr, qOpts).Optimize(e.logicalOptimizers)
@@ -260,6 +270,9 @@ func (e *compatibilityEngine) NewRangeQuery(ctx context.Context, q storage.Query
 		ExtLookbackDelta:         e.extLookbackDelta,
 		EnableAnalysis:           e.enableAnalysis,
 		NoStepSubqueryIntervalFn: e.noStepSubqueryIntervalFn,
+	}
+	if qOpts.StepsBatch > 64 {
+		return nil, ErrStepsBatchTooLarge
 	}
 
 	lplan, warns := logicalplan.New(expr, qOpts).Optimize(e.logicalOptimizers)
@@ -410,11 +423,7 @@ loop:
 	var result parser.Value
 	switch q.expr.Type() {
 	case parser.ValueTypeMatrix:
-		matrix := promql.Matrix(series)
-		if matrix.ContainsSameLabelset() {
-			return newErrResult(ret, extlabels.ErrDuplicateLabelSet)
-		}
-		result = matrix
+		result = promql.Matrix(series)
 	case parser.ValueTypeVector:
 		// Convert matrix with one value per series into vector.
 		vector := make(promql.Vector, 0, len(resultSeries))
@@ -439,9 +448,6 @@ loop:
 			}
 		}
 		sort.Slice(vector, q.resultSort.comparer(&vector))
-		if vector.ContainsSameLabelset() {
-			return newErrResult(ret, extlabels.ErrDuplicateLabelSet)
-		}
 		result = vector
 	case parser.ValueTypeScalar:
 		v := math.NaN()