--- a/engine/enginefuzz_test.go
+++ b/engine/enginefuzz_test.go
@@ -443,7 +443,7 @@ func getSeries(ctx context.Context, q storage.Queryable) ([]labels.Labels, error
 func validateTestCases(t *testing.T, cases []*testCase) {
 	failures := 0
 	for i, c := range cases {
-		if !cmp.Equal(c.oldRes, c.newRes, comparerFuzzing) {
+		if !cmp.Equal(c.oldRes, c.newRes, comparer) {
 			for _, load := range c.loads {
 				t.Logf(load)
 			}