* check for duplicate labels at every step
* binary operations bail out early if high card side is empty