@MichaHoffmann 2024-01-07
> Can we check bench results

Can do in a bit, but ultimately we need to check for duplicates at every step since thats what prometheus is doing too roughly.