@MichaHoffmann 2024-01-08
> LGTM overall, I just wonder if we need to wrap the user defined operator too.

I think generally yes since we have no preconditions to prove otherwise!